third_party/WebKit/LayoutTests/editing/pasteboard/dataTransfer-setData-getData-expected.txt

Tests correct behavior of event.dataTransfer.setData/getData for 'URL', 'text/uri-list' and 'text/plain'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

--- Test set/get 'URL':
PASS getDataResultType is "string"
PASS getDataResult is "http://test.com/"
--- Test set/get 'URL' with multiple URLs:
PASS getDataResultType is "string"
PASS getDataResult is "http://test.com/"
--- Test set/get 'text/uri-list':
PASS getDataResultType is "string"
PASS getDataResult is "http://test.com\r\nhttp://check.com"
--- Test set/get 'text/uri-list' using '\n':
PASS getDataResultType is "string"
PASS getDataResult is "http://test.com\nhttp://check.com"
--- Test set 'text/uri-list', get 'URL':
PASS getDataResultType is "string"
PASS getDataResult is "http://test.com/"
--- Test set 'URL', get 'text/uri-list':
PASS getDataResultType is "string"
PASS getDataResult is "http://test.com\r\nhttp://check.com"
--- Test set 'text/uri-list', get 'URL', using only '\n':
PASS getDataResultType is "string"
PASS getDataResult is "http://test.com/"
--- Test set/get 'text/uri-list' with comments:
PASS getDataResultType is "string"
PASS getDataResult is "# comment\r\nhttp://test.com\r\nhttp://check.com"
--- Test set 'text/uri-list', get 'URL' with comments:
PASS getDataResultType is "string"
PASS getDataResult is "http://test.com/"
--- Test set 'text/uri-list', get 'URL' with only comments:
PASS getDataResultType is "string"
PASS getDataResult is ""
--- Test set/get 'text/plain':
PASS getDataResultType is "string"
PASS getDataResult is "Lorem ipsum dolor sit amet."
--- Test set 'text/plain', get 'text':
PASS getDataResultType is "string"
PASS getDataResult is "Lorem ipsum dolor sit amet."
--- Test set 'text', get 'text/plain':
PASS getDataResultType is "string"
PASS getDataResult is "Lorem ipsum dolor sit amet."
--- Test set/get 'text':
PASS getDataResultType is "string"
PASS getDataResult is "Lorem ipsum dolor sit amet."
PASS successfullyParsed is true

TEST COMPLETE
